+++ Nudge/UI/Common/InformationButton.swift
@@ -54,8 +54,7 @@ struct InformationButtonAsAction: View {
     
     var body: some View {
         Button(action: {
-            UIUtilities().openMoreInfoUnsupported()
-            UIUtilities().postUpdateDeviceActions(userClicked: true, unsupportedUI: true)
+            UIUtilities().updateDeviceUnsupported()
         }) {
             Text(.init(UserInterfaceVariables.actionButtonTextUnsupported.localized(desiredLanguage: getDesiredLanguage(locale: appState.locale))))
         }